Card to display track information
A component is created to display track information. The component
is imported in the ITunesSearch Container and it takes in a track
as props. The card displays track name, artist and there is also a
simple player that provides the preview for the particular track.
The trackCard component is created using floki.