detailed track card
An ITunesProvider container is created using floki. This houses a TrackContainer,
to display a list of tracks to the user based on the search query and a TrackDetailsContainer,
that displays details of specific track when the user clicks 'See More' button on the track.
The TrackDetailsContainer first tries to fetch the track information from the store, if the data
doesn't exist in the store, the container makes an API call and fetches the data. Three Card Actions are
provided, which link to the arsists's page, albums' page and tracks' page respectively.